refactor symbol tests
