+++ Source/components/__tests__/SymbolInstance.test.js
@@ -1,0 +1,39 @@
+/* globals expect, test */
+/* eslint-disable no-param-reassign */
+
+import { SymbolMaster, Text, Artboard } from '../../'
+
+function createSymbolMaster(document) {
+  const artboard = new Artboard({
+    name: 'Test',
+    parent: document.selectedPage,
+  })
+  const text = new Text({
+    text: 'Test value',
+    parent: artboard,
+  })
+
+  // build the symbol master
+  return {
+    master: SymbolMaster.fromArtboard(artboard),
+    text,
+    artboard,
+  }
+}
+
+test('should have overrides', (context, document) => {
+  const { master, text } = createSymbolMaster(document)
+  const instance = master.createNewInstance()
+  document.selectedPage.layers = document.selectedPage.layers.concat(instance)
+
+  expect(instance.overrides.length).toBe(1)
+  const override = instance.overrides[0]
+  expect(override.toJSON()).toEqual({
+    id: `${text.id}_stringValue`,
+    path: text.id,
+    property: 'stringValue',
+    symbolOverride: false,
+    value: 'Test value',
+    isDefault: true,
+  })
+})
